# feat(html/codegen): improve perf