--- a/crates/swc_html_codegen/src/lib.rs
+++ b/crates/swc_html_codegen/src/lib.rs
@@ -804,13 +804,23 @@ where
             attribute.push('=');
 
             if self.config.minify {
-                let minifier = minify_attribute_value(value, self.quotes);
+                let (minifier, quote) = minify_attribute_value(value, self.quotes);
+
+                if let Some(quote) = quote {
+                    attribute.push(quote);
+                }
 
                 attribute.push_str(&minifier);
-            } else {
-                let normalized = normalize_attribute_value(value);
 
+                if let Some(quote) = quote {
+                    attribute.push(quote);
+                }
+            } else {
+                let normalized = escape_string(value, true);
+
+                attribute.push('"');
                 attribute.push_str(&normalized);
+                attribute.push('"');
             }
         }
 
@@ -820,15 +830,11 @@ where
     #[emitter]
     fn emit_text(&mut self, n: &Text) -> Result {
         if self.ctx.need_escape_text {
-            let mut data = String::with_capacity(n.data.len());
-
             if self.config.minify {
-                data.push_str(&minify_text(&n.data));
+                write_multiline_raw!(self, n.span, &minify_text(&n.data));
             } else {
-                data.push_str(&escape_string(&n.data, false));
+                write_multiline_raw!(self, n.span, &escape_string(&n.data, false));
             }
-
-            write_multiline_raw!(self, n.span, &data);
         } else {
             write_multiline_raw!(self, n.span, &n.data);
         }
@@ -927,9 +933,20 @@ where
 }
 
 #[allow(clippy::unused_peekable)]
-fn minify_attribute_value(value: &str, quotes: bool) -> String {
+fn minify_attribute_value(value: &str, quotes: bool) -> (Cow<'_, str>, Option<char>) {
     if value.is_empty() {
-        return "\"\"".to_string();
+        return (Cow::Borrowed(value), Some('"'));
+    }
+
+    // Fast-path
+    if !quotes
+        && value.chars().all(|c| match c {
+            '&' | '`' | '=' | '<' | '>' | '"' | '\'' => false,
+            c if c.is_ascii_whitespace() => false,
+            _ => true,
+        })
+    {
+        return (Cow::Borrowed(value), None);
     }
 
     let mut minified = String::with_capacity(value.len());
@@ -969,32 +986,18 @@ fn minify_attribute_value(value: &str, quotes: bool) -> String {
     }
 
     if !quotes && unquoted {
-        return minified;
+        return (Cow::Owned(minified), None);
     }
 
     if dq > sq {
-        format!("'{}'", minified.replace('\'', "&apos;"))
+        return (Cow::Owned(minified.replace('\'', "&apos;")), Some('\''));
     } else {
-        format!("\"{}\"", minified.replace('"', "&quot;"))
+        return (Cow::Owned(minified.replace('"', "&quot;")), Some('"'));
     }
-}
-
-fn normalize_attribute_value(value: &str) -> String {
-    if value.is_empty() {
-        return "\"\"".to_string();
-    }
-
-    let mut normalized = String::with_capacity(value.len() + 2);
-
-    normalized.push('"');
-    normalized.push_str(&escape_string(value, true));
-    normalized.push('"');
-
-    normalized
 }
 
 #[allow(clippy::unused_peekable)]
-fn minify_text<'a>(value: &'a str) -> Cow<'a, str> {
+fn minify_text(value: &str) -> Cow<'_, str> {
     // Fast-path
     if value.is_empty() {
         return Cow::Borrowed(value);
@@ -1128,7 +1131,7 @@ fn minify_amp(chars: &mut Peekable<Chars>) -> String {
 // 4. If the algorithm was not invoked in the attribute mode, replace any
 // occurrences of the "<" character by the string "&lt;", and any occurrences of
 // the ">" character by the string "&gt;".
-fn escape_string<'a>(value: &'a str, is_attribute_mode: bool) -> Cow<'a, str> {
+fn escape_string(value: &str, is_attribute_mode: bool) -> Cow<'_, str> {
     // Fast-path
     if value.is_empty() {
         return Cow::Borrowed(value);