<!-- Note: CI script will automatically rebase your PR so please do not rebase unless required -->

<!-- Thanks for submitting a pull request! Please provide enough information so that others can review your pull request. -->

**Description:**

Performance:

```
html/minify/document/css_spec                                                                           
                        time:   [38.763 ms 38.821 ms 38.915 ms]
                        change: [-4.1127% -3.6128% -3.0759%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) high mild
html/minify/document/github                                                                           
                        time:   [16.572 ms 16.694 ms 16.862 ms]
                        change: [-3.3237% -2.7138% -1.7989%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) high mild
html/minify/document/stackoverflow                                                                           
                        time:   [14.690 ms 15.023 ms 15.315 ms]
                        change: [-3.6187% -2.1485% -0.4165%] (p = 0.02 < 0.05)
                        Change within noise threshold.
Found 1 outliers among 10 measurements (10.00%)

```

**BREAKING CHANGE:**

No

**Related issue (if exists):**

No